Remove `@dnd` based Dashboard component
