+++ packages/frappe-ui-react/package.json
@@ -39,8 +39,6 @@
   },
   "dependencies": {
     "@base-ui/react": "^1.1.0",
-    "@dnd-kit/core": "^6.3.1",
-    "@dnd-kit/sortable": "^10.0.0",
     "@floating-ui/react": "^0.27.16",
     "@headlessui/react": "^2.2.9",
     "@popperjs/core": "^2.11.8",
