+++ packages/frappe-ui-react/src/components/dashboard/dashboardWidgetGalleryItem.tsx
