+++ packages/frappe-ui-react/src/components/dashboard/index.ts
@@ -1,2 +1,5 @@
-export { Dashboard, default } from './dashboard';
+export { Dashboard, default } from "./dashboard";
+export { WidgetWrapper } from "./widgetWrapper";
+export { DashboardWidgetGallery } from "./dashboardWidgetGallery";
+export { DashboardProvider } from "./dashboardContext";
 export * from "./types";
