--- packages/frappe-ui-react/src/components/dashboard/layoutContext.tsx
@@ -1,11 +1,0 @@
-/**
- * External dependencies.
- */
-import { createContext } from "react";
-
-/**
- * Internal dependencies.
- */
-import { LayoutContextValue } from "./types";
-
-export const LayoutContext = createContext<LayoutContextValue | null>(null);
