--- packages/frappe-ui-react/src/components/dashboardGrid/index.ts
@@ -1,5 +1,0 @@
-export { DashboardGrid, default } from './dashboard';
-export { WidgetWrapper } from './widgetWrapper';
-export { DashboardWidgetGallery } from './dashboardWidgetGallery';
-export { DashboardProvider } from './dashboardContext';
-export * from "./types";
